+++ sdk/storage/Azure.Storage.Blobs/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 12.9.0-beta.4 (Unreleased)
 - Added support for service version 2020-08-04.
 - Added support for Blob Query Parquet input format.
+- Added WithCustomerProvidedKey() and WithEncryptionScope() to BlobClient, BlobBaseClient, AppendBlobClient, and PageBlobClient.
 - Fixed bug where clients would sometimes throw a NullReferenceException when calling GenerateSas() with a BlobSasBuilder parameter.
 - Fixed bug where BlobBaseClient.Exists() would not function correctly on blobs encrypted with CPK.